--- tests/utils/__init__.py
@@ -1,1 +1,0 @@
-from tests.utils.base_test_case import *
